separate friendrequests page to two components
